+++ Sources/EudiWalletKit/Services/Openid4VpUtils.swift
@@ -36,12 +36,6 @@ class OpenId4VpUtils {
 	// example path: $.given_name_national_character
 	static let pathItemRx: NSRegularExpression = try! NSRegularExpression(pattern: #"\$\.(.+)"#, options: .caseInsensitive)
 
-	static func generateSessionTranscript(clientId: String,	responseUri: String, nonce: String, jwkThumbprint: String? = nil) -> SessionTranscript {
-		let openID4VPHandover = generateOpenId4VpHandover(clientId: clientId, responseUri: responseUri,	nonce: nonce, jwkThumbprint: jwkThumbprint)
-		return SessionTranscript(handOver: openID4VPHandover)
-	}
-
-
 	/// Generate the `SessionTranscript` CBOR where the presentation request is invoked using redirects.
 	/// - Parameters:
 	///   - clientId: The `client_id` request parameter. If applicable, this includes the Client Identifier Prefix.
@@ -50,18 +44,10 @@ class OpenId4VpUtils {
 	///   - jwkThumbprint: If the response is encrypted, e.g., using `direct_post.jwt`, this element must be the JWK SHA-256 Thumbprint of the Verifier's public key used to encrypt the response. Otherwise `nil`.
 	/// - Returns: A CBOR representation of the OpenID4VPHandover structure.
 	/// - Remark: See [Handover and SessionTranscript Definitions](https://openid.net/specs/openid-4-verifiable-presentations-1_0.html#name-handover-and-sessiontranscript).
-	static func generateOpenId4VpHandover(clientId: String,	responseUri: String, nonce: String, jwkThumbprint: String? = nil) -> CBOR {
-		var openID4VPHandoverInfo: [UInt8]
-
-		if let jwkThumbprint = jwkThumbprint {
-			let openID4VPHandoverInfoToHash = CBOR.array([.utf8String(clientId), .utf8String(nonce), .utf8String(jwkThumbprint), .utf8String(responseUri)])
-			openID4VPHandoverInfo = [UInt8](SHA256.hash(data: openID4VPHandoverInfoToHash.asData()))
-		}
-		else {
-			let openID4VPHandoverInfoToHash = CBOR.array([.utf8String(clientId), .utf8String(nonce), .null, .utf8String(responseUri)])
-			openID4VPHandoverInfo = [UInt8](SHA256.hash(data: openID4VPHandoverInfoToHash.asData()))
-		}
-
+	static func generateOpenId4VpHandover(clientId: String,	responseUri: String, nonce: String, jwkThumbprint: [UInt8]? = nil) -> CBOR {
+		let jwkThumbprintCbor: CBOR = jwkThumbprint != nil ? .byteString(jwkThumbprint!) : .null
+		let openID4VPHandoverInfoToHash = CBOR.array([.utf8String(clientId), .utf8String(nonce), jwkThumbprintCbor, .utf8String(responseUri)])
+		let	openID4VPHandoverInfo = [UInt8](SHA256.hash(data: openID4VPHandoverInfoToHash.asData()))
 		return CBOR.array(["OpenID4VPHandover", .byteString(openID4VPHandoverInfo)])
 	}
 
@@ -121,7 +107,7 @@ class OpenId4VpUtils {
 		guard let allPathsDict = (try sdJwt.recreateClaims()).disclosuresPerClaimPath else { throw WalletError(description: "No disclosures found") }
 		let allPaths = Array(allPathsDict.keys)
 		let query = Set(allPaths.filter { path in requestItems.contains(where: { r in r.claimPath.contains2(path) }) })
-		let presentedSdJwt = try await sdJwt.present(query: query)
+		let presentedSdJwt = try sdJwt.present(query: query)
 		guard let presentedSdJwt else { return nil }
 		let digestCreator = DigestCreator(hashingAlgorithm: hashingAlg)
 		guard let sdHash = digestCreator.hashAndBase64Encode(input: CompactSerialiser(signedSDJWT: presentedSdJwt).serialised) else { return nil }
@@ -164,18 +150,6 @@ extension ClaimPathElement {
 		if case .claim(let name) = self { name } else if case .arrayElement(let index) = self { String(index) } else { "" }
 	}
 }
-
-extension CoseEcCurve {
-	init?(crvName: String) {
-		switch crvName {
-		case "P-256": self = .P256
-		case "P-384": self = .P384
-		case "P-512": self = .P521
-		default: return nil
-		}
-	}
-}
-
 
 extension CredentialQuery {
 	public var docType: String? {